moved provider interface out of internal
